Refactor: Modularize solar system simulator codebase
This commit refactors the solar system simulator by splitting the monolithic `main.js` file into smaller, more manageable modules based on functionality. This improves maintainability and extensibility.

The following modules have been created in a new `js/` directory:
- `dom.js`: Centralizes DOM element lookups.
- `utils.js`: Contains utility functions for scaling.
- `data.js`: Stores the `planetData` array.
- `rings.js`: Handles the creation of planetary rings.
- `moons.js`: Handles the creation of moons.
- `scene.js`: Manages the Three.js scene, camera, renderer, and controls.
- `interactions.js`: Manages user interactions like mouse clicks and the speed slider.

Additionally, this commit adds a fallback mechanism for asset loading. If the texture for Saturn's rings fails to load, a non-textured, colored ring will be rendered in its place, and a warning will be logged to the console.
