+++ js/dom.js
@@ -1,0 +1,7 @@
+export const infoPanel = document.getElementById('info-panel');
+export const infoName = document.getElementById('info-name');
+export const infoRadius = document.getElementById('info-radius');
+export const infoDistance = document.getElementById('info-distance');
+export const infoPeriod = document.getElementById('info-period');
+export const speedSlider = document.getElementById('speed-slider');
+export const canvas = document.querySelector('#bg');
